Moving parse stack trace functions over to new file
